Begin reading PAP files
This lays the groundwork for actually reading PAP files, but the Havok
data part isn't parsed yet.

+++ src/envb.rs
@@ -37,7 +37,7 @@ impl ReadableFile for Envb {
     fn from_existing(platform: Platform, buffer: ByteSpan) -> Option<Self> {
         let endianness = platform.endianness();
         let mut cursor = Cursor::new(buffer);
-        let string_heap = StringHeap::from(cursor.position());
+        let string_heap = StringHeap::from(cursor.position() as i64);
 
         Envb::read_options(&mut cursor, endianness, (&string_heap,)).ok()
     }
