+++ src/essb.rs
@@ -37,7 +37,7 @@ impl ReadableFile for Essb {
     fn from_existing(platform: Platform, buffer: ByteSpan) -> Option<Self> {
         let endianness = platform.endianness();
         let mut cursor = Cursor::new(buffer);
-        let string_heap = StringHeap::from(cursor.position());
+        let string_heap = StringHeap::from(cursor.position() as i64);
 
         Essb::read_options(&mut cursor, endianness, (&string_heap,)).ok()
     }
