+++ src/lvb.rs
@@ -43,7 +43,7 @@ pub struct Lvb {
 impl ReadableFile for Lvb {
     fn from_existing(platform: Platform, buffer: ByteSpan) -> Option<Self> {
         let mut cursor = Cursor::new(buffer);
-        let string_heap = StringHeap::from(cursor.position());
+        let string_heap = StringHeap::from(cursor.position() as i64);
 
         Lvb::read_options(&mut cursor, platform.endianness(), (&string_heap,)).ok()
     }
