+++ src/sgb.rs
@@ -31,7 +31,7 @@ impl ReadableFile for Sgb {
     fn from_existing(platform: Platform, buffer: ByteSpan) -> Option<Self> {
         let endianness = platform.endianness();
         let mut cursor = Cursor::new(buffer);
-        let string_heap = StringHeap::from(cursor.position());
+        let string_heap = StringHeap::from(cursor.position() as i64);
 
         Sgb::read_options(&mut cursor, endianness, (&string_heap,)).ok()
     }
